third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/w3c/webperf/submission/Google/resource-timing/html/test_resource_iframe_self_navigation-expected.txt

This is a testharness.js-based test.
PASS http://127.0.0.1:8000/w3c/webperf/resources/inject_resource_test.html is expected to have initiatorType iframe 
Harness: the test ran to completion.
